fix(freebsd): incorrect constant type from ttycom.h

ref: https://github.com/rust-lang/libc/pull/493

the value and type of the constant are obtained by executing the following script

ttycom.sh:
```
#/bin/sh

cat <<EOF
#include <stdio.h>
#include <sys/ttycom.h>
#include <sys/time.h>
#include <termios.h>

void main() {
EOF

grep -o TIOC[A-Z_]* /usr/include/sys/ttycom.h | sort -u |\
 xargs -I{} echo '    printf("pub const {}: u%lu = 0x%lx;\n", sizeof({})*8, {});'

echo '}'
```

```
./ttycom.sh > main.c
clang main.c
./a.out
```

+++ src/unix/bsd/freebsdlike/freebsd/mod.rs
@@ -2158,9 +2158,9 @@ pub const CTL_P1003_1B_SEM_VALUE_MAX: ::c_int = 23;
 pub const CTL_P1003_1B_SIGQUEUE_MAX: ::c_int = 24;
 pub const CTL_P1003_1B_TIMER_MAX: ::c_int = 25;
 
-pub const TIOCGPTN: ::c_uint = 0x4004740f;
-pub const TIOCPTMASTER: ::c_uint = 0x2000741c;
-pub const TIOCSIG: ::c_uint = 0x2004745f;
+pub const TIOCGPTN: ::c_ulong = 0x4004740f;
+pub const TIOCPTMASTER: ::c_ulong = 0x2000741c;
+pub const TIOCSIG: ::c_ulong = 0x2004745f;
 pub const TIOCM_DCD: ::c_int = 0x40;
 pub const H4DISC: ::c_int = 0x7;
